Fix 'getSetupTemplate' so it accepts some custom user configuration in 'chartOptions'
* Custom axis configuration example in `test/examples/dataviz/c3/index.html`
* Do not write over x axis configurations if they exist in getSetupTemplate
* Add tests for `getSetupTemplate`
* Pull `getSetupTemplate` and `getFormatDateDefault` into separate file for easier testing
