+++ test/unit/index.js
@@ -13,3 +13,5 @@ require("./modules/core/tracker/browser-spec");
 require("./modules/core/request/browser-spec");
 require("./modules/dataviz/dataviz-spec");
 require("./modules/dataviz/parseRawData-spec");
+
+require("./modules/dataviz/adapters/get-setup-template-test");
